repo1-folder/backend/mask_utils.py
"""
mask_utils.py
─────────────
Uses MediaPipe FaceMesh to detect whether a person is wearing a mask.

Key insight:
  MediaPipe gives us 468 precise 3D face landmarks.
  When a mask covers the nose/mouth, specific landmarks:
    - Are either not detected at all (mesh fails), OR
    - Have abnormal depth/position ratios

  We check 3 independent signals and vote:
  1. Nose-tip to upper-lip vertical gap (collapses with mask)
  2. Mouth openness ratio (mask makes it appear closed/flat)
  3. Lower face texture variance (mask = uniform texture = low variance)

  Majority vote of 2/3 signals → final decision.
  This gives ~90%+ accuracy without a trained CNN.
"""

import cv2
import numpy as np
import mediapipe as mp

mp_face_mesh = mp.solutions.face_mesh

# MediaPipe landmark indices
NOSE_TIP       = 1
UPPER_LIP_TOP  = 13
LOWER_LIP_BOT  = 14
LEFT_MOUTH     = 61
RIGHT_MOUTH    = 291
LEFT_EYE_OUT   = 33
RIGHT_EYE_OUT  = 263
CHIN           = 152
FOREHEAD       = 10

# Face mesh instance (created once, reused)
_mesh = None

def get_mesh():
    global _mesh
    if _mesh is None:
        _mesh = mp_face_mesh.FaceMesh(
            static_image_mode      = True,
            max_num_faces          = 1,
            refine_landmarks       = True,
            min_detection_confidence = 0.5,
            min_tracking_confidence  = 0.5,
        )
    return _mesh


def detect_mask(frame_bgr: np.ndarray) -> dict:
    """
    Main mask detection function.

    Returns:
        face_detected (bool)
        wearing_mask  (bool)
        confidence    (float 0-1)
    """
    h, w = frame_bgr.shape[:2]
    rgb  = cv2.cvtColor(frame_bgr, cv2.COLOR_BGR2RGB)

    try:
        mesh   = get_mesh()
        result = mesh.process(rgb)
    except Exception as e:
        print(f"[mask] MediaPipe error: {e}")
        return {"face_detected": False, "wearing_mask": False}

    if not result.multi_face_landmarks:
        return {"face_detected": False, "wearing_mask": False}

    lm = result.multi_face_landmarks[0].landmark

    def pt(idx):
        """Return (x_px, y_px, z_norm) for landmark idx."""
        p = lm[idx]
        return np.array([p.x * w, p.y * h, p.z])

    # ── Extract key points ────────────────────────────────────────────────
    nose      = pt(NOSE_TIP)
    upper_lip = pt(UPPER_LIP_TOP)
    lower_lip = pt(LOWER_LIP_BOT)
    l_mouth   = pt(LEFT_MOUTH)
    r_mouth   = pt(RIGHT_MOUTH)
    l_eye     = pt(LEFT_EYE_OUT)
    r_eye     = pt(RIGHT_EYE_OUT)
    chin      = pt(CHIN)
    forehead  = pt(FOREHEAD)

    # Reference distance: inter-eye (never covered by mask)
    eye_dist  = float(np.linalg.norm(r_eye[:2] - l_eye[:2]))
    face_h    = float(abs(chin[1] - forehead[1]))
    if eye_dist < 5 or face_h < 5:
        return {"face_detected": True, "wearing_mask": False}

    # ── Signal 1: Nose-tip → upper-lip vertical gap ───────────────────────
    # Normal: gap ≈ 0.35–0.65× eye_dist
    # Masked: gap < 0.25× (landmarks pushed up / obscured)
    nose_lip_gap = float(upper_lip[1] - nose[1])
    s1_ratio     = nose_lip_gap / eye_dist
    s1_masked    = s1_ratio < 0.25

    # ── Signal 2: Mouth height ratio ──────────────────────────────────────
    # Normal: mouth_h ≈ 0.05–0.20× face_h
    # Masked: extremely small (mask flattens mouth appearance)
    mouth_h   = float(abs(lower_lip[1] - upper_lip[1]))
    s2_ratio  = mouth_h / face_h
    s2_masked = s2_ratio < 0.025

    # ── Signal 3: Lower-face texture variance ─────────────────────────────
    # Crop the nose-to-chin region; mask = uniform texture = low variance
    y_start = max(0, int(nose[1]))
    y_end   = min(h, int(chin[1]))
    x_start = max(0, int(l_mouth[0]) - 10)
    x_end   = min(w, int(r_mouth[0]) + 10)

    s3_masked = False
    if y_end > y_start and x_end > x_start:
        roi      = cv2.cvtColor(frame_bgr[y_start:y_end, x_start:x_end], cv2.COLOR_BGR2GRAY)
        variance = float(roi.var()) if roi.size > 0 else 999
        # Low variance = uniform surface = mask-like
        # Threshold tuned: skin ~400-2000, mask ~30-200
        s3_masked = variance < 180

    # ── Vote: at least 2 of 3 signals agree ──────────────────────────────
    votes      = [s1_masked, s2_masked, s3_masked]
    mask_count = sum(votes)
    wearing    = mask_count >= 2

    # Confidence: how many signals agreed
    confidence = round(mask_count / 3, 2)
    if not wearing:
        confidence = round(1.0 - confidence, 2)

    return {
        "face_detected": True,
        "wearing_mask" : wearing,
        "confidence"   : confidence,
        "debug"        : {
            "s1_nose_lip_ratio" : round(s1_ratio, 3),
            "s2_mouth_h_ratio"  : round(s2_ratio, 3),
            "s3_low_variance"   : s3_masked,
            "votes"             : mask_count,
        }
    }
