repo1-folder/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
python-docx==1.1.2
Pillow==10.3.0
deepface==0.0.93
tensorflow==2.15.0
tf-keras==2.15.1
mediapipe==0.10.14
